fix: update quantity handling in save_prod_receipt method for correct data type

+++ app/Http/Controllers/Backend/TransactionController.php
@@ -910,11 +910,11 @@ class TransactionController extends Controller
 
             $lines        = [];
             foreach ($validated['stocks'] as $row) {
-                $entryQty = (float) str_replace(',', '.', str_replace('.', '', $row['qty']));                // Qty baru yang diinput
+                $entryQty = (float) str_replace(',', '.', str_replace('.', '', $row['qty']));
                 // untuk API SAP
                 $lines[] = [
                     'BaseEntry'    => (int) $row['BaseEntry'],
-                    'Quantity'    =>  (int)  $entryQty,
+                    'Quantity'    => $entryQty,
                     'WhsCode'    =>  $warehouse ?? '',
                 ];
